+++ packages/client/signers-cryptography/src/algorithms/asymmetric/ecdh.ts
@@ -1,0 +1,7 @@
+import { deriveSymmetricKey } from "../../primitives";
+
+export class ECDH {
+    async deriveSharedSecret(privateKey: CryptoKey, publicKey: CryptoKey): Promise<CryptoKey> {
+        return deriveSymmetricKey(privateKey, publicKey);
+    }
+}