+++ packages/wallets/src/wallets/wallet-factory.ts
@@ -5,9 +5,11 @@ import type { Chain } from "../chains/chains";
 import type { InternalSignerConfig, SignerConfigForChain, ExternalWalletInternalSignerConfig } from "../signers/types";
 import { Wallet } from "./wallet";
 import { assembleSigner } from "../signers";
-import type { WalletArgsFor } from "./types";
-import { EmailSigner } from "@/signers/email/email";
+import type { WalletArgsFor, WalletOptions } from "./types";
 import { deepCompare } from "@/utils/signer-validation";
+import { EvmEmailSigner } from "@/signers/email/evm-email-signer";
+import { SolanaEmailSigner } from "@/signers/email/solana-email-signer";
+import { isEVMBlockchain } from "@crossmint/common-sdk-base";
 
 export class WalletFactory {
     constructor(private readonly apiClient: ApiClient) {}
@@ -75,7 +77,7 @@ export class WalletFactory {
         walletResponse: GetWalletSuccessResponse,
         args: WalletArgsFor<C>
     ): Wallet<C> {
-        const signerConfig = this.toInternalSignerConfig(walletResponse, args.signer);
+        const signerConfig = this.toInternalSignerConfig(walletResponse, args.signer, args.options);
         return new Wallet(
             {
                 chain: args.chain,
@@ -89,7 +91,8 @@ export class WalletFactory {
 
     private toInternalSignerConfig<C extends Chain>(
         walletResponse: GetWalletSuccessResponse,
-        signer: SignerConfigForChain<C>
+        signer: SignerConfigForChain<C>,
+        options?: WalletOptions
     ): InternalSignerConfig<C> {
         if (signer == null) {
             throw new WalletCreationError("Signer is required to create a wallet");
@@ -137,8 +140,12 @@ export class WalletFactory {
 
             case "email": {
                 if (
-                    walletResponse.type !== "solana-smart-wallet" ||
-                    walletResponse.config.adminSigner.type !== "solana-keypair"
+                    !(
+                        (walletResponse.type === "solana-smart-wallet" &&
+                            walletResponse.config.adminSigner.type === "solana-keypair") ||
+                        (walletResponse.type === "evm-smart-wallet" &&
+                            walletResponse.config.adminSigner.type === "evm-keypair")
+                    )
                 ) {
                     throw new WalletCreationError("Wallet signer 'email' has no address");
                 }
@@ -151,7 +158,7 @@ export class WalletFactory {
                     signerAddress: address,
                     crossmint: this.apiClient.crossmint,
                     onAuthRequired: signer.onAuthRequired,
-                    _handshakeParent: signer._handshakeParent,
+                    clientTEEConnection: options?.clientTEEConnection,
                 };
             }
 
@@ -221,15 +228,20 @@ export class WalletFactory {
             if (email == null) {
                 throw new Error("Email is required to create a wallet with email signer");
             }
-            if (chain !== "solana") {
-                throw new Error("Email signer is only supported for Solana wallets");
-            }
 
-            const emailSignerAddress = await EmailSigner.pregenerateSigner(email, this.apiClient.crossmint);
-            return {
-                type: "solana-keypair",
-                address: emailSignerAddress,
-            };
+            if (chain === "solana") {
+                const emailSignerAddress = await SolanaEmailSigner.pregenerateSigner(email, this.apiClient.crossmint);
+                return {
+                    type: "solana-keypair",
+                    address: emailSignerAddress,
+                };
+            } else if (isEVMBlockchain(chain)) {
+                const emailSignerAddress = await EvmEmailSigner.pregenerateSigner(email, this.apiClient.crossmint);
+                return {
+                    type: "evm-keypair",
+                    address: emailSignerAddress,
+                };
+            }
         }
 
         throw new Error("Invalid signer type");