+++ apps/payments/create-react-app/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @crossmint/client-sdk-react-ui-starter
 
+## 1.3.111
+
+### Patch Changes
+
+- @crossmint/client-sdk-react-ui@2.2.7
+
+## 1.3.110
+
+### Patch Changes
+
+- Updated dependencies [2ab0922]
+- Updated dependencies [9b8220a]
+  - @crossmint/client-sdk-react-ui@2.2.6
+
+## 1.3.109
+
+### Patch Changes
+
+- @crossmint/client-sdk-react-ui@2.2.5
+
 ## 1.3.108
 
 ### Patch Changes