+++ apps/wallets/smart-wallet/next/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @crossmint/client-sdk-smart-wallet-next-starter
 
+## 0.1.139
+
+### Patch Changes
+
+- @crossmint/client-sdk-react-ui@2.2.7
+
+## 0.1.138
+
+### Patch Changes
+
+- Updated dependencies [2ab0922]
+- Updated dependencies [9b8220a]
+  - @crossmint/client-sdk-react-ui@2.2.6
+
+## 0.1.137
+
+### Patch Changes
+
+- @crossmint/client-sdk-react-ui@2.2.5
+
 ## 0.1.136
 
 ### Patch Changes