+++ apps/wallets/smart-wallet/react/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @crossmint/client-sdk-smart-wallet-react-starter
 
+## 1.5.60
+
+### Patch Changes
+
+- @crossmint/client-sdk-smart-wallet@0.3.20
+- @crossmint/client-sdk-smart-wallet-web3auth-adapter@0.2.23
+
+## 1.5.59
+
+### Patch Changes
+
+- @crossmint/client-sdk-smart-wallet@0.3.19
+- @crossmint/client-sdk-smart-wallet-web3auth-adapter@0.2.22
+
 ## 1.5.58
 
 ### Patch Changes