+++ packages/client/react-base/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @crossmint/client-sdk-react-base
 
+## 0.5.7
+
+### Patch Changes
+
+- 6a342c4: Move \_handshakeParent from signer to options config and rename to clientTEEConnection
+- dabb284: Fix createOnLogin wallet types based on chain
+- Updated dependencies [daaf328]
+- Updated dependencies [6a342c4]
+- Updated dependencies [dabb284]
+  - @crossmint/wallets-sdk@0.10.7
+  - @crossmint/common-sdk-base@0.8.2
+  - @crossmint/client-sdk-base@1.5.8
+
+## 0.5.6
+
+### Patch Changes
+
+- Updated dependencies [583757c]
+  - @crossmint/wallets-sdk@0.10.6
+
+## 0.5.5
+
+### Patch Changes
+
+- Updated dependencies [38abd83]
+- Updated dependencies [38abd83]
+- Updated dependencies [ba8495c]
+  - @crossmint/wallets-sdk@0.10.5
+  - @crossmint/common-sdk-base@0.8.1
+  - @crossmint/client-signers@0.0.18
+  - @crossmint/client-sdk-base@1.5.7
+
 ## 0.5.4
 
 ### Patch Changes