+++ packages/client/ui/vanilla-ui/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @crossmint/client-sdk-vanilla-ui
 
+## 1.1.54
+
+### Patch Changes
+
+- Updated dependencies [dabb284]
+  - @crossmint/common-sdk-base@0.8.2
+  - @crossmint/client-sdk-base@1.5.8
+
+## 1.1.53
+
+### Patch Changes
+
+- Updated dependencies [38abd83]
+  - @crossmint/common-sdk-base@0.8.1
+  - @crossmint/client-sdk-base@1.5.7
+
 ## 1.1.52
 
 ### Patch Changes