+++ packages/client/ui/vue-ui/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @crossmint/client-sdk-vue-ui
 
+## 1.1.54
+
+### Patch Changes
+
+- @crossmint/client-sdk-base@1.5.8
+
+## 1.1.53
+
+### Patch Changes
+
+- @crossmint/client-sdk-base@1.5.7
+
 ## 1.1.52
 
 ### Patch Changes