+++ packages/client/wallets/smart-wallet/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @crossmint/client-sdk-smart-wallet
 
+## 0.3.20
+
+### Patch Changes
+
+- Updated dependencies [dabb284]
+  - @crossmint/common-sdk-base@0.8.2
+  - @crossmint/client-sdk-base@1.5.8
+
+## 0.3.19
+
+### Patch Changes
+
+- Updated dependencies [38abd83]
+  - @crossmint/common-sdk-base@0.8.1
+  - @crossmint/client-sdk-base@1.5.7
+
 ## 0.3.18
 
 ### Patch Changes