+++ packages/client/wallets/walletconnect/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @crossmint/client-sdk-walletconnect
 
+## 1.2.19
+
+### Patch Changes
+
+- Updated dependencies [dabb284]
+  - @crossmint/common-sdk-base@0.8.2
+
+## 1.2.18
+
+### Patch Changes
+
+- Updated dependencies [38abd83]
+  - @crossmint/common-sdk-base@0.8.1
+
 ## 1.2.17
 
 ### Patch Changes