+++ packages/common/base/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @crossmint/common-sdk-base
 
+## 0.8.2
+
+### Patch Changes
+
+- dabb284: Fix createOnLogin wallet types based on chain
+
+## 0.8.1
+
+### Patch Changes
+
+- 38abd83: Fixed issue with jwt not being set correctly
+
 ## 0.8.0
 
 ### Minor Changes