+++ packages/wallets/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @crossmint/wallets-sdk
 
+## 0.10.7
+
+### Patch Changes
+
+- daaf328: Add EVM signing
+- 6a342c4: Move \_handshakeParent from signer to options config and rename to clientTEEConnection
+- dabb284: Fix createOnLogin wallet types based on chain
+- Updated dependencies [dabb284]
+  - @crossmint/common-sdk-base@0.8.2
+  - @crossmint/common-sdk-auth@1.0.39
+
+## 0.10.6
+
+### Patch Changes
+
+- 583757c: Add EVM email signer
+
+## 0.10.5
+
+### Patch Changes
+
+- 38abd83: Fixed issue with email not set error even when email is set
+- Updated dependencies [38abd83]
+- Updated dependencies [ba8495c]
+  - @crossmint/common-sdk-base@0.8.1
+  - @crossmint/client-signers@0.0.18
+  - @crossmint/common-sdk-auth@1.0.38
+
 ## 0.10.4
 
 ### Patch Changes